To do: mostrar nombre de producto en historial de pedidos

+++ database/migrations/2024_05_11_162811_create_pedidos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             // $table->foreignId('carrito_id')->nullable()->constrained()->on('carritos')->nullOnDelete();
             $table->foreignId('producto_id')->nullable()->constrained()->on('productos')->nullOnDelete();
             $table->foreignId('direccion_id')->nullable()->constrained()->on('direcciones')->nullOnDelete();
-            $table->integer('total');
+            $table->decimal('total', 8, 2);
             $table->enum('estado', ['pendiente', 'enviado', 'entregado', 'cancelado']);
             $table->timestamps();
         });
